The describe input function is mandatory
Array data source now implements the describe() input function
and the higher level data source handler and expression handler
code relies on the describe function to be present.

United the two follower arrays in ocrpt_array_test to show the
effect and ordering of the plain follower and N:1 follower queries.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>