Add more optimization opportunities
Moved newblankexpr() to exprutil.c from ocrptgrammar.y and made
it independent of the grammar. The details necessary for grammar
handling is in newblankexpr1() in ocrptgrammar.y.

The grammar now makes a note in ocrpt_expr about surrounding
parentheses.

Introduced small space saving changes: ocrpt_expr.type and
other struct members are now bitfields instead of full size int,
bool or enum. Same for ocrpt_result and ocrpt_function struct
members.

Added the "left-associative" marker to ocrpt_function.

For commutative and associative functions, the operands can be
reordered. This allows pre-computing more constant expressions.

For commutative and associative functions, if all operands are of
the same function type, the subexpression operands can be pulled
up to the parent level, eliminating subexpressions.

For left-associative functions, the same subexpression elimination
can be done freely for the first subexpression, but for subsequent
subexpressions it can be done if it's not inside parentheses.
This is is purely theoretical, as the grammar creates the expression
tree according to the left- or right-associativity.

Added new expressions to ocrpt_expr_test to exercise the optimizer
changes.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>