Add new function required to implement report breaks
The new function is called ocrpt_expr_cmp_results(), returns
true if the two data rows in the same expression is identical.

They are not considered identical if:
* any of the two ocrpt_result structure pointers is not set
* they are of different types, e.g. one numeric and one error,
  caused by conversion issues
* they are NULLs (SQL behaviour)
* they are errors, regardless of the error message

Otherwise the values themselves are checked.

Modified the array data source to initialize results with
isnull=true so the new function returns true for the first row.

Modified the array data source API functions to indicate the
correct type of the array data.

Extended ocrpt_array_test with a new numeric field called "adult"
with "yes"/"no" values to test the new function and adapted to
the call interface changes.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>