Implement val() function and string concatenation via add()
The val() function converts a string to number. It also converts
boolean values like "true", "false", "yes", "no" and for making it
easier for PostgreSQL, "t" and "f" as 1 and 0, respectively.

val() is a NOP when fed with a numeric value. the result in this
case equals the operand value.

Modify ocrpt_mem_string_new_with_len() so the passed-in length
is simply the allocated length and the passed-in string can be
shorter.

Implemented ocrpt_mem_string_resize() so re-allocate the string
storage of an ocrpt_string structure.

String concatenation in function add() used the above changes
in ocrpt_string handling as an optimized case to reduce the
number of allocations during datasource traversal.

Extended ocrpt_expr_test to exercise the new additions.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>