Allow optimizing 0-operand functions
Remove the protection in ocrpt_expr_optimize_worker() against
NULL e->ops because it's normal for 0-operand functions.

Introduce a new flag in struct ocrpt_function called dont_optimize
and use it in ocrpt_expr_optimize_worker() to optimize function
calls with constants that are allowed to be optimized. Currently,
random() is the only one not allowed to be optimized.

This allows e.g. abs(nulln()) to get evaluated to constant NULL
expression of numeric type in ocrpt_expr_optimize().

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>