Implement beginnings of report descriptor XML parsing
Added a DTD file (still not complete) based on rlib.dtd but
extended with capabilities to set up data sources and queries
straight from the XML. Also, the addition of a new toplevel
XML node was necessary to describe the possibility of using
either <Report> or <Part> as the toplevel node in the report.

The currently parsed and processed XML nodes are:
* <OpenCReport> as the toplevel node
* <Datasources> as the parent node for <Datasource> nodes
* <Datasource> as a description for one datasource
* <Queries> as the parent node for <Query> nodes
* <Query> as a decription of one query

New API additions:
* ocrpt_query_find() to find the query using its name,
  similarly to ocrpt_datasource_find()
* ocrpt_parse_xml() to parse and process an XML file

The internal struct symbol "ocrpt_array_input" was made public
across the library source files, so a datasource ->input pointer
can be checked against it to determine the type of the datasource.

A new test program "ocrpt_array_xml_test" was added to exercise
the new XML parser. The supplementary arrayquery.xml file contains
the information about the datasource and query used by the application.
The application must be compiled with -rdynamic, just like as
the previously added ocrpt_array_resolve_test.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>