Implement the CSV datasource
Unconditionally require libcsv to provide this functionality.
New APIs:
* ocrpt_datasource_add_csv() to add a CSV datasource filter
* ocrpt_query_add_csv() to add a query using a CSV file

ocrpt_query_add_csv() opens and parses the file passed to it
immediately and creates the array in memory to be used by
the array datasource functions. It can also use a similar
supplementary array for the field types as the array type,
the fields are not guessed and considered to be strings if
the supplementary "coltypes" array is not passed by the
application.

As a performance helper fix, modified ocrpt_list_end_append()
interface so the last element in the list is an in-out parameter.

Added two new test programs: ocrpt_csv_test and ocrpt_csv_array_test
to exercise the CSV datasource and query by itself and a mixed case
of a CSV datasource/query and an ARRAY datasource/query as its
follower in both variants. ocrpt_csv_array_test is a variant of
ocrpt_array_test where the followers were first tested on two array
queries.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>