Implement JSON datasource and query handling
OpenCReports now requires the YAJL headers and library.
The format of the JSON string accepted as a query is:

{
 "columns": [ column names as array of strings ],
 "coltypes": [ column types as array of strings],
 "rows": [
  { "column": "value", ... },
  ...
 ]
}

The presence of the "coltypes" array is optional and can be
supplied by the application via the ocrpt_query_add_json()
API function similarly to the array and CSV queries.
Accepted strings in "coltypes" are: "string", "number" and
"datetime". When the coltypes array is present in JSON and is
also passed by the application, the JSON array is used.

If "coltypes" is present, then it should have the same number of
elements or fewer than the "columns" array. Excess elements in
"coltypes" are ignored. Missing elements at the end of the
"coltypes" array are implicitly "string" types.

The column names in the object elements in the "rows" array
are validated against the column names. Missing columns will
be NULL. When a column name is duplicated, the last one wins.
Unknown column names are ignored.

XML handling of JSON datasources and queries are also added.

Add test programs to exercise certain aspects of JSON queries:
* ocrpt_json_test, "coltypes" is passed by the application,
  no error in JSON
* ocrpt_json2_test, "coltypes" is both passed by the application
  and its array is present in the JSON (the JSON array wins),
  no error in JSON
* ocrpt_json3_test, "coltypes" is present in the JSON, NULL is
  passed by the application, no error in JSON
* ocrpt_json4_test, the "coltypes" array has fewer elements than
  "columns" (the missing type specifier is "string")
* ocrpt_json5_test, the "coltypes" array has more elements than
  "columns" (excess is ignored), one row has a missing element
  (will be NULL in the query row), one row has a duplicate element
  (the last one wins)
* ocrpt_json_xml_test, same as ocrpt_json_test but uses report XML
  where "coltypes" is specified and the array is present in the
  application
* ocrpt_json_xml2_test, same as ocrpt_json3_test but used report XML
  and "coltypes" is not supplied by the XML, nor the application

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>