Implement report XML expression fallbacks and add the PostgreSQL query API call
<Datasource> and <Query> tag attributes and the <Query> tag text
element now all can be written as is (plain text) or as an
OpenCReport expression. For example, name="some_name" and
name="'some_name'" are both accepted and will use the "some_name"
string value. This can be done because the expression parsing
and resolution now exclude runtime variable references, so
the expression evaluation returns an error. The error can be
detected and the string value can fall back to the original XML
text value.

Also added ocrpt_query_add_postgresql(), which is a dummy function
for now.

Added two more PostgreSQL test cases to exercise unquoted strings
and settings in the report XML.

Also fixed ocrpt_array_xml_test because the report XML was modified
to it was calling ocrpt_datasource_find() with the wrong name.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>