Test circular or conflicting follower queries
The leader and follower queries must be different. The same query
cannot be a regular and N:1 follower at the same time. Regular and
N:1 follower queries must not form a cyclic graph.

The query result array was leaked in case query navigation was
not started. Fixed the memory leak.

Added a test case exercising disallowed followers.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>