Add rownum() function
Add a rownum() function with an optional query name parameter.
rownum() without a parameter will use the main query,
rownum('queryname') will use the named query.

The result is the current row number (1-based) into the query.

In case the named query is a follower and it runs out of rows,
or in case of a 1:N follower and the condition match comes up
empty in the follower query, the result is a numeric NULL.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>