Make sure breaks really	trigger	on the first row
If the query's last two rows happened to match the first row in
a way that the break field expression's results matched, then
the break did not trigger on the first row upon re-running the query.

Make sure breaks do trigger on the first row.

Added a new unit test called ocrpt_break_xml3_test to exercise this.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>