Implement break row number accounting
Implemented break row number accounting using a self-referencing
iterative expression.

Added a brrownum("breakname") function that can be used both
internally and by reports.

As a micro-optimization, an ocrpt_break pointer was added to
ocrpt_expr which is set at the first evaluation, used directly
the next time.

Added a new unit test called ocrpt_break_rownum_xml_test to
exercise the brrownum() function.

The same micro-optimization was done for the rownum() function:
an ocrpt_query pointer was added to ocrpt_expr. Adapted the
unit test called ocrpt_rownum_test because this change caused
a use-after-free in the test which re-used the same expression
pointer across freeing and re-creating the ocrpt_query structure
which was used by the expression via rownum('queryname').

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>