Fix clang-analyzer reports in parsexml.c
They were harmless unused variables but I like the code to compile
with no warnings even with static analyzers.

There are a few potential NULL pointer dereference warnings coming
from lex.yy.c but it's a generated file and the reports are not
in the code spelled out in the symbolic ocrptpatterns.l so there's
nothing to be done about this and likely are false positives.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>