Don't remove an expression the "r->exprs" list when the report is executing
It isn't added to the r->exprs list in ocrpt_expr_parse() either
when the report is executing, so make it symmetric.

And also make it faster for the case when an expression is parsed,
evaluated and freed in a custom function during running the report.
This is a performance fix for emulating RLIB's rlib_compile_infix()
PHP function.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>