Add RLIB compatibility mode
New API: ocrpt_set_rlib_compat()

For now, this only modifies the behaviour of the ^ operator.

By default ^ uses the xor() function, i.e. the operator behaves
as in the C language.

When RLIB compatibility mode is set, ^ uses the pow() function,
i.e. it behaves as in the BASIC language.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>