Add dtos() function
Add missing break; to ocrpt_expr_init_result_internal().

Add dtos() function for expression to return the date part of
a datetime as string. If the date part is not valid, the result
is NULL. An interval operand is invalid.

Extend the ocrpt_datetime_expr_test unit test to print a set
of examples in different locales.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>