Add date() and now() functions
These provide the current date and current timestamp as in
being the start time of executing the report, or if used in
expressions without a report, the date and time of the first
execution of these functions.

Added an unstable unit test called ocrpt_now_test to exercise
the new functions.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>