Implement year(), month(), day() and dim() functions
Implemented year(), month(), day() functions. These functions
take a datetime operand with valid date part or an interval,
and return the corresponding number, i.e. the year, month or
day-of-month, respectively. They return NULL for a datetime
without a valid date part.

Implemented dim() function. This function only accepts a
datetime with valid date part but not an interval. This is
because an interval contains offset values and not an actual
date, so the days-in-month value cannot be interpreted in a
meaningful way.

All of them return NULL if the operand is NULL.

Extended the ocrpt_datetime_expr_test unit test to cover these.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>