Add wiy(), wiyo(), wiy1() and stdwiy() functions
Added wiy(), wiyo(), wiy1() and stdwiy() functions. These return
the week-of-year value for a datetime with valid date part.

wiy() considers the first Sunday in the year as the start of
the first week.

wiyo() uses a numeric weekday offset (Monday = 1, Tuesday = 2, ...)
to consider as the as the start of the first week.

wiy1() considers the first Monday in the year as the start of the
first week. It may be considered as a shortcut for wiyo(d, 1).

stdwiy() uses the ISO-8601 notation about week numbers. Week 1 is
the first week that has at least 4 days in the new year.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>