Implement chgdateof() and chgtimeof() functions
Implemented chgdateof() and chgtimeof() functions.
They take two datetime (not interval) operands.
The result is the first operand modified with the date
or the time part of the second operand, respectively.

If the date part of the second operand is not valid for
chgdateof(), it will set to not not valid in the result.

If the time part of the second operand is not valid for
chgtimeof(), it will set to not not valid in the result.

If the result has no valid date and time parts, then it
will be NULL.

Extended the ocrpt_datetime_expr_test unit test to cover
these cases.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>