Implement gettimeinsecs() and settimeinsecs() functions
Implemented gettimeinsecs() and settimeinsecs() functions.

gettimeinsecs() takes a datetime value and returns its time part
in seconds if valid, NULL otherwise.

settimeinsecs() takes two operands: a datetime value and a number,
and changes the datetime value's time part to the value of the
second operand interpreted as seconds from 0:00:00.

Initialize the datetime members in nulldt() to all zero and
set the date and time parts to not valid.

Extended the ocrpt_datetime_expr_test unit test to cover these.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>