Fix a memory leak in the expression parser
Fixed a memory leak in the expression parser for cases when two
or more distinct and valid expression strings are concatenated
together but which are not valid as one expression.

Also initialized yyextra in the expression parser. Testing with
AddressSanitizer revealed that it may contain garbage that
caused problems in the parse error paths.

Accidentally noticed while testing other expressions.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>