Add str() function
str(op1, op2, op3) takes op1 as a number and converts it
to a string of op2 character length with op3 number of decimal
places. The actual string may be longer than op2 if the number
is larger and takes more characters to be printed.

This is for RLIB compatibility.

Adapted the unit test ocrpt_expr_test.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>