Add printf() formatting function
It accepts arbitrary number of arguments, the first argument is
the format string just like in C. The rest are data that must
be the same type as described in the format string.

Fixed printing datetime in new style formats with brackets.

Adapted the unit test ocrpt_expr_format_test to exercise printf()
and show the datetime fix.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>