Move NULL check to ocrpt_format_string() from ocrpt_format() and ocrpt_dtosf()
This way, when the format string has literals around the format
specifier, the literals are printed.

It also fixes a crash when the printf() function is used in
report expressions with NULL data.

Adapted the unit test ocrpt_expr_format_test to exercise NULLs
and NULLs with the wrong data type for the format specifier.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>