Modify ocrpt_part_append_report() API
Modified ocrpt_part_append_report() API interface to return an
ocrpt_part * pointer.

Since the passed-in ocrpt_part * pointer may be NULL, a new
ocrpt_part structure is allocated and the report is added to it.
Return the pointer to newly allocated structure in this case.

In case the passed-in ocrpt_part * pointer was valid, return it.

Use this in parsexml.c for standalone <Report> XML sections
without a parent <Part>. Some tags for <Report> (that are currently
not parsed) are duplicates of those applicable to the parent <Part>
section and should be applied to the parent <Part> section in this
case.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>