Allow passing unquoted strings for align="..."
The words "left", "right", "center" can be identifiers (i.e. query
columns) as far as the expression parser is concerned.

Check for these words first then try to parse the attribute as
an expression.

Also remove a debugging call to ocrpt_expr_print() on the line
element value. Adapt one unit test result because this output
ended up at the end of the PDF stream.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>