Handle report result values in the output with formatting
Use either the explicit format="..." attribute, or default formats
and format values with ocrpt_format_string() when printing in pdf.c

The formatted string is stored in an ocrpt_string structure instead
of separate char * and length members in struct ocrpt_line_element.
Account for this change when freeing the structure to avoid leaking
memory.

Added a new unit test to exercise one report with query data and:
* <NoData> which is not shown because there is data
* <ReportHeader> and <ReportFooter>
* <Detail> with both <FieldHeaders> and <FieldDetails>

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>