Handle multi-page reports
Handle multi-page reports by passing the page height to the
layout code and split calculating an output section height
from drawing the section even more. This way, when a section
to be printed on the page reaches the bottom margin (including
the height of the optional page footer) then print the section
on a new page.

Print print page header and footer properly with correct values
for r.pageno (current page number) and r.totpages (total number
of pages).

Added two unit test to exercise multi-page PDF reports:
* layout_pdf_xml6_test with field_header_priority="low"
* layout_pdf_xml7_test with default field_header_priority.

layout_pdf_xml7_test caught a bug with r.pageno on the last page.
Data traversal must backtrack one row so the pre-evaluated
expressions for r.pageno and r.totpages are valid when printed.

suppressPageHeaderFirstPage now sets the flag for <Part> even if
it's set for <Report>. This is a convenience function inherited
from RLIB, so report XMLs with only a <Report> section without
the encapsulating <Part> section can set the page header, footer
and this flag for the implicitly created <Part> parent.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>