PDF layout fixes
When the width is not specified for a <literal> or a <field>
element and it's the only element in a <Line>, use the whole
page width assigned for the report. This allows the alignment
to work intuitively, i.e. the text is aligned along the
whole page width.

When the width is not specified for a <literal> or a <field>
element and it's NOT the only element in a <Line>, then calculate
the field width as the text length of the element (counted in
UTF-8 characters) multiplied by the font width used for the parent
<Line>.

Don't truncate text. Instead, use a bounding box (masking) to
clip the text visually when it would not fit into the field
width. This makes the PDF contain the complete text so it can
be copied and pasted elsewhere and allows not special casing
alignments.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>