Implement multi-part report rows
Implemented multi-part report rows, a.k.a. handling multiple <pd>
elements in <pr>.

Page number and page vertical position is saved when the first
<pd> element is started, so subsequent ones can start from the
same page and position.

<pd> elements without a width specifier are sharing the total
page width.

Empty <pd> elements with no <Report> element(s) inside may
serve as a vertical separator if its width is specified.

Attributes "detail_columns" and "column_pad" are now specific
to <pd> instead of <Report>. The reasoning is that it's not
well defined to have different number of columns for different
<Report> pieces in the same <pd>. "detail_columns" and
"column_pad" attributes may still be used in a <Report> element
as shortcuts if the XML toplevel element is the <Report> but
they are ignored when a <pd> parent element is spelled out
in the report XML.

Same for the margin attributes: these settings belong to
the <Part> element but they can still be used as shortcuts
if the XML toplevel element is <Report>.

Fixed a small memory leak for <load> elements: the canonical
filename was not freed.

Fixed creating the report using the <load> element: all
the parent <Part>, <pr> and <pd> elements are defined with
a backing structure pointer for each, so use them when parsing
the loaded <Report>. This fixes a confusion about where the
report element belongs.

Added a new unit test called layout_pdf_xml17_test that exercises
multiple <pd> elements, including one empty <pd> with no <Report>
as a separator.

Also fixed some -Wextra warnings (especially unused function
parameters) for some internal APIs and signed vs unsigned
comparison warnings in functions.c

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>