Implement height handling for <pd> and PDF links
<pd> elements may have an optional width attribute.

Handle it in a way that if the report output would exceed
the specified width, fields on the right side of the width
boundary are not printed in a row and field that would be
split by the width boundary are partially masked with a
bounding box.

Data is still processed as usual so delayed expressions
and precalculated variables would be shown correctly for
the remaining visible fields.

Added URL link handling to the PDF output.
<field> and <literal> elements can use the "link" attribute
to specify the URL.

Added two unit test:
* layout_pdf_xml24_test to exercise PDF links in <field> and
  <literal> elements
* layout_pdf_xml25_test to exercise <pd width="..."> for a
  two-column report

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>