Handle <pd> with height and page break in the middle
Handle <pd> with height and page break in the middle in a way
that the remaining height would flow to the next page.
The repeated field header obviously takes away space on the
next page. This is a corner case so it's the report's designer's
responsibility.

Fix tracking the start and end of a <pd> element, so
* other <pd> elements in the same <pr> group can return
  to the same starting point
* <pd> elements of the next <pr> group start where the longest
  <pd> of the current <pr> group ended.

Added two new unit tests called layout_pdf_xml26_test and
layout_pdf_xml27_test to exercise the new details.

No changes in other unit tests.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>