Handle suppressPageHeaderFirstPage for <Part> and <Report>
Handle suppressPageHeaderFirstPage for the <Part>, i.e. the
<Pageheader> node output can be suppressed on the first page.

When multiple <Part> elements are in the same report XML, this
property is only applied to the first <Part> since these elements
(and iterations of them) always start on a new page. Subsequent
report <Part> elements therefore cannot start on the first page,
so the page header suppression on the first page cannot apply.

When <Report> is the toplevel XML node or is without a parent
<Part> node, the attribute suppressPageHeaderFirstPage is
a shortcut. The property is set for the implicitly created
parent <Part> node.

Added a new unit test called layout_pdf_xml33_test to exercise
suppressPageHeaderFirstPage.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>