Allow <Image> as a child element in <Line>
Handle new attributes for <Image>:
* "textWidth" (also accepted as "text_width) - RLIB had this
  it's the line element width designated for the image
* "bgcolor" - the background color for the image
* "align" - alignment for the image when it's used as
  a line element

The last two attributes do not exist in RLIB.

The background color may be useful because the image file
may be an SVG and the outline of the image is not filled.
Or it may be a raster image format with transparency.
If the image is used as a line element, the parent <Line>
may have the background color set. In this case, the same
background should be applied to the image.

Extended the documentation with the new feature.

Added a new unit test called layout_pdf_xml41_test to exercise
an image as a line element, including  alignment and background
color.

Adapted unit test results which now have minute differences
compared to their previous results:
* layout_pdf_xml3_test
* layout_pdf_xml32_test
* layout_pdf_xml34_test
* layout_pdf_xml37_test

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>