Add a new PHP unit test
Add a new PHP unit test called layout_pdf_translate_xml1_test.php
which is a rewritten copy of the C unit test calles
layout_pdf_translate_xml1_test.c. The expected result of the
former is the identical copy of the expected result of the latter.

Tweak tests/Makefile.am so the translation tests are automatically
included in "make pdf-test" and "make php-pdf-test" if translation
tests are enabled.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>