Try to fix a crash after using OpenCReport\Query::get_result()
Rename object pointer names consistently for better grepping.

Check return value of ocrpt_query_get_result() in
OpenCReport\Query::get_result()

Set reference pointers to NULL when freeing a
OpenCReport\QueryResult object.

Since all of the above was not enough to avoid a crash in
a new PHP unit test with a typo when PHP was exiting and
freeing objects: add a hack to test for pointer validity and
use it on dubious pointers. I got 0x10 instead of NULL
on a pointer that was set to NULL beforehand. It seems the
PHP runtime poisons data. (?)

The typo in the unit test was using $q->get_result($param)
where $q->get_result() expects zero parameters. The exception
was correctly thrown, but PHP was crashing upon exit.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>