Keep the original type of query columns
The original type is supplied by either the database source
or the supplemental column type array for array or file based
sources.

Keep that in the query result structure so it is always used as
the base type, from which auto-conversion may deviate to
an error type.

No change in unit test results.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>