Modify the PHP API
Change OpenCReport::expr_parse() to return OpenCReport\Expr
or NULL, instead of an array. The error string can be read
via OpenCReport::expr_error().

Adapted previously existing unit tests.

Fixed php_opencreport_query_discover_array() so it returns NULL
and 0 for width/height for non-existing or not-asked-for arrays.

Modified OpenCReport\Result::get_string() to return the string
value (or NULL) even the value is not a string. It mimics the
C API better this way and the unit tests expect this semantics.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>