Automatically convert unresolvable field references to strings
Automatically convert unresolvable field references to strings
in ocrpt_expr_resolve() and optionally warn when it happens.

This will allow debugging reports better, because the field
value will be the field name as is instead of becoming empty.

Suppress warnings for parsing datasource and query names from
report XMLs because it's already documented that datasource and
query names can be plain strings, despite being parsed as an
expression. This behaviour is expected in RLIB, too.

Adapted the result of the C unit test header_query_test because
it used a nonexisting field, in part intentionally.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>