Allow <Report> specific <PageHeader> and <PageFooter>
Relax the previous notion that <PageHeader> and <PageFooter>
in single <Report>s are shortcuts for spelling out the whole
XML structure:
  <Part>
    <PageHeader>...</PageHeader>
    <PageFooter>...</PageFooter>
    <pr>
      <pd>
        <Report>...</Report>
      </pd>
    </pr>
  </Part>

Instead, make <PageHeader> and <PageFooter> in <Report>
use the ocrpt_report structure pointer, in order to be able
to use report specific expressions, allowing more interesting
use cases.

New warnings are added if multiple <PageHeader> and <PageFooter>
sections are found in the same <Part> or in its child <Report>
nodes. Only the first <PageHeader> and <PageFooter> are used,
subsequent ones are ignored just like before. This was previously
silent.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>