Render page header at the start of rendering the page
Now that page headers and footers may contain report expressions,
it is more natural to render the page header at the start of
rendering the page.

This way the page header gets values based on the first query row
on the page instead of values based on the last query row.

Page footers are still rendered at the end of rendering the page.

This caused miniscule rendering differences in several PDF
unit tests that are not visible to the naked eye, but these
tests still failed because ImageMagick's compare command can
detect the differences.

Re-generate the expected results for these tests.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>