Allow size_unit to be set from a query
Move evaluating size_unit="..." from parsexml.c to executing the
report. The former way mandated that this setting must have been
a constant. The new way allows it to be set from a supplementary
query.

ocrpt_set_size_unit_points() was renamed to ocrpt_set_size_unit()
and changed its prototype, so an expression is passed instead of
a bool constant.

Adapted documentation.

As expected, 3 unit test results changed that now emit an
invalid field reference warning for either "points" or "rlib"
and they are converted to string constants.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>