Allow part page orientation to be set from a query
Move evaluating orientation="..." from parsexml.c to executing the
report. The former way mandated that this setting must have been
a constant. The new way allows it to be set from a supplementary
query.

ocrpt_part_set_landscape() was renamed to ocrpt_part_set_orientation()
and changed its prototype, so an expression is passed instead of
a bool constant.

Adapted documentation.

2 unit test were adapted to the API change with no result changes.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>