Rename layout C API arguments to "expr_string"
Renamed the remaining C API functions' arguments to "expr_string"
to reflect the fact that they are expression strings and not plain
string parameters. The modified functions are:
* ocrpt_part_set_font_name()
* ocrpt_part_set_font_size()
* ocrpt_part_set_paper_by_name()
* ocrpt_part_set_top_margin()
* ocrpt_part_set_bottom_margin()
* ocrpt_part_set_left_margin()
* ocrpt_part_set_right_margin()

Also, add fallback to use the parameter as a string in case
expression parsing fails for ocrpt_part_set_paper_by_name().
It is likely a NOP, as paper names ("A4", etc.) may be valid
identifiers.

Adapted documentation.

No change in unit test results.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>