Rewrite documentation generation
After adding the PHP API documentation, the amount of input
overwhelmed docbook2pdf and the PDF was generated erroneously.

Switch to using xmllint and xsltproc to generate the HTML
documentation from the SGML sources. Also use fop to generate
the documentation in PDF, now in A4 and US letter formats.
Use a subset of the XML style sheets from the PostgreSQL project
for professional formatting.

Adapt SGML files to satisfy xmllint and the style sheets.

Generating documentation is much faster now.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>