Rename Hungarian accented field name
Generating the PDF documentation now throws this warning and replaces
"ő" with "#" in the identifier name, which is not a good identifier
after that:

WARNING: Glyph "ő" (0x151, odblacute) not available in font "Courier".

Renamed the Hungarian accented field name to avoid the warning
and the character replacement for the missing glyph.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>