Set the realloc and free functions for the CSV parser
Add a non-const variant declaration of ocrpt_mem_free_t
which matches the free function declaration of libcsv.

Use it for casting ocrpt_mem_free0 to the necessary function
type to avoid a compiler warning.

Using the same allocator for libcsv as other parts of the
library helps making the opencreport.so PHP module friendler
to PHP's memory_limit setting.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>