Add the XML output driver
Modified struct ocrpt_output_functions to satisfy the XML driver:
* added ocrpt_break pointer to several existing member functions
* removed start_no_data and end_no_data member functions
  (it turned out to be not necessary)

Adapted callers and existing output drivers to the changes.

Added the XML output driver, which uses libXML2 internally to
create the XML tree and dump it in text form at the end.
The output format attempts to mimic the RLIB XML output format.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>