Preparation for the TXT output driver
Add a flag to the common private data to indicate whether an
output driver can use line element specific font name and size.

Set the flag for the PDF, HTML and XML drivers.

This is for preparing for the TXT output driver that can't use
line element fonts. In this regard, this will make the TXT output
driver more closely mimic the RLIB behaviour.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>