Add TXT output driver
Added a new boolean flag to struct ocrpt_output_functions called
"support_any_font" - the TXT driver does not deal with fonts.

When the support_any_font flag is not set, the layout code
used Courier font and 12 point font size, but still computes
the font width and fields widths according to the font width
when the width calculation is done using the RLIB compatible
method.

When this flag is not set, the rendered fields don't get
truncated beyond the page or column width. This also fixes
one CSV unit test result.

The TXT output driver does not set the new flag and also doesn't
support page breaks or multi-column layout.

Due to the implicitly used 12 point font size, the layout is
rendered differently than other drivers that do adhere to the
font sizes set in the report XML or from code. It is especially
noticable for memo (multi-line) fields.

Added unit test results for the TXT driver.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>