Handle the eval() function through grammar transformations
It was a previous oversight to omit the eval() RLIB function
that evaluates an embedded expression string.

Handle this function using grammar transformation if:
* there is no "eval" user function defined
* there is only a single argument
* the passed-in argument is a string constant

When the above conditions are fulfilled, the embedded expression
is parsed and inserted into the currently parsed expression tree
in place of the eval() function call. This makes the embedded
expression subject to optimizations.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>