Add strlen() function
Add the last missing expression function that existed in RLIB.

The strlen() function (obviously) returns the number of characters
in the string.

But since OpenCReports exclusively uses UTF-8, the result is the
number of UTF-8 characters, as opposed to RLIB that basically
had a single-byte encoding mindset, so it returned the number of
bytes. This fits better with other string arithmetics functions,
like left(), right() and mid().

Extended the expr_test C and PHP unit tests.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>