Fix use-after-free in the argument list handling in the parser
The recently added vector comparison handling added code to better
handle multiple argument lists in the same expression using a list
of lists structure, but forgot to add protection against adding a
the same list twice on that list of lists.

Any reasonably complex expression with multiple function calls
reveals a use-after-free bug in this code, of which many are in
the unit tests already. But I forgot to run the whole test suite
using ASAN/UBSAN before releasing 0.8.1. ¯\_(ツ)_/¯

Add the protection against duplicate entries on the same list.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>