Use a better workaround for the PHP module crash at shutdown
Instead of waiting 1 second at module shutdown, perform a dummy
PDF report run at module startup, which is very fast and only
slows down module startup by a small fraction of a second so
PHP unit tests are fast again.

The original reasoning for adding the 1 second delay at shutdown
in commit f10bbbe ("Add 1 second sleep for PHP module shutdown")
was:

This is suspected to be a threaded library initialization
in one of the libraries used by OpenCReports that does not
finish by the time PHP finishes and the OpenCReports PHP module
gets unloaded. The library thread crashes because PHP calls
dlclose(). Observed on Fedora 38 and Fedora 39.

The dummy PDF report run warms up the rendering libraries (Cairo
and Pango) so they can synchronize and wait for their own
initializations to finish behind the scenes.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>